Split backend media storage boundary
