+++ apps/backend/src/mediaAssets/storage/presigned.test.ts
@@ -1,0 +1,80 @@
+import assert from "node:assert/strict";
+import test from "node:test";
+import {
+  createPresignedMediaAssetUploadPartsWithDependencies,
+  createPresignedMediaAssetUploadWithDependencies,
+} from ".";
+import {
+  createTestS3Client,
+  getTestMediaAssetsStorageConfig,
+  testLastOperationId,
+  testLastOperationIdSha256,
+  testMediaAssetId,
+  testObservationScope,
+  testStagingStorageKey,
+  testUploadStorageKey,
+  testWorkspaceId,
+} from "./testHelpers";
+
+test("createPresignedMediaAssetUploadWithDependencies signs all returned required upload headers", async () => {
+  const upload = await createPresignedMediaAssetUploadWithDependencies(
+    {
+      workspaceId: testWorkspaceId,
+      mediaAssetId: testMediaAssetId,
+      storageKey: testUploadStorageKey,
+      mimeType: "image/png",
+      sha256: "a".repeat(64),
+      lastOperationId: testLastOperationId,
+      observationScope: testObservationScope,
+    },
+    {
+      s3Client: createTestS3Client(),
+      getMediaAssetsStorageConfigFn: getTestMediaAssetsStorageConfig,
+    },
+  );
+
+  const signedHeaders = new URL(upload.url).searchParams.get("X-Amz-SignedHeaders");
+  assert.notEqual(signedHeaders, null);
+  const signedHeaderSet = new Set(signedHeaders?.split(";") ?? []);
+  for (const headerName of Object.keys(upload.headers)) {
+    assert.ok(signedHeaderSet.has(headerName), `Expected ${headerName} to be signed`);
+  }
+  assert.equal(upload.headers["x-amz-meta-flashcards-workspace-id"], testWorkspaceId);
+  assert.equal(upload.headers["x-amz-meta-flashcards-media-asset-id"], testMediaAssetId);
+  assert.equal(upload.headers["x-amz-meta-flashcards-sha256"], "a".repeat(64));
+  assert.equal(upload.headers["x-amz-meta-flashcards-last-operation-id-sha256"], testLastOperationIdSha256);
+});
+
+test("createPresignedMediaAssetUploadPartsWithDependencies signs per-part checksum headers", async () => {
+  const partUrls = await createPresignedMediaAssetUploadPartsWithDependencies(
+    {
+      workspaceId: testWorkspaceId,
+      mediaAssetId: testMediaAssetId,
+      stagingStorageKey: testStagingStorageKey,
+      s3UploadId: "s3-upload-id-1",
+      parts: [
+        {
+          partNumber: 1,
+          sha256: "a".repeat(64),
+        },
+      ],
+      observationScope: testObservationScope,
+    },
+    {
+      s3Client: createTestS3Client(),
+      getMediaAssetsStorageConfigFn: getTestMediaAssetsStorageConfig,
+    },
+  );
+
+  const partUrl = partUrls[0];
+  assert.notEqual(partUrl, undefined);
+  assert.equal(partUrl?.method, "PUT");
+  assert.equal(partUrl?.partNumber, 1);
+  assert.equal(
+    partUrl?.headers["x-amz-checksum-sha256"],
+    Buffer.from("a".repeat(64), "hex").toString("base64"),
+  );
+  const signedHeaders = new URL(partUrl?.url ?? "").searchParams.get("X-Amz-SignedHeaders");
+  assert.notEqual(signedHeaders, null);
+  assert.ok(new Set(signedHeaders?.split(";") ?? []).has("x-amz-checksum-sha256"));
+});
